prevent app initialization on testing
