+++ backend/package.json
@@ -6,8 +6,8 @@
         "start": "node --experimental-specifier-resolution=node dist/src/app.js",
         "dev": "ts-node src/app.ts",
         "build": "tsc -p . --outDir ./dist",
-        "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles --watchAll --coverage",
-        "test:ci": "node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles",
+        "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --watchAll --coverage",
+        "test:ci": "node --experimental-vm-modules node_modules/jest/bin/jest.js",
         "format": "prettier --write --list-different .",
         "format:check": "prettier --check ."
     },
